+++ src/main/java/ovdm/bus/LunchMoney.java
@@ -1,8 +1,8 @@
 package ovdm.bus;
 
 public class LunchMoney extends Money implements ILunchProvisionable {
-    @Override public int countItems() { return 0; }
-    @Override public int countItems(String type) { return 0; }
+    @Override public long countItems() { return 0; }
+    @Override public long countItems(String type) { return 0; }
 
     @Override
     public double moneyAmount() {
